Create README.md
Using the Shiny app on R the attached code illustrates the accidental deaths in the US between 1973 and 1978. From selecting the month in the given drop down menu (Jan-Dec) a resulting line graph will automatically appear illustrating the amount of deaths in the particular month across the years of 1973-1978. This is a helpful way to quickly identify the year/month in which the most accidental deaths had occurred. All data was provided through R. This data can be accessed and viewed by the following: View(USAccDeaths)